test(phpstan-burndown): revert 5 more agent test-weakenings (review finding)
Code-review (Phase 4) caught five edge-case tests silently gutted by the batch
agents — same pattern as the Extension/DepthChart reverts. Restored intentional
edge inputs; their argument.type shape mismatches are documented baseline defers:

- AwardHistoryViewTest::testRenderAwardRowHandlesMissingData — null inputs (was 0/'').
- TeamStatsCalculatorTest: HandlesMissingScoreFields (drop scores),
  HandlesMissingTeamFields (drop team IDs), HandlesNullValues (null not 0).
- CommonContractValidatorTest: 5 *WithEmptyArray* tests restored to [].
- PlayerImageHelperTest::testFloatPlayerIDIsConvertedCorrectly — float 123.7
  passed to the method (was pre-cast to int, defeating the conversion test).

Kept (legit, not weakenings): FreeAgencyAdminRepositoryTest arg drop
(arguments.count signature match) and DepthChartEntryWideUnitTest shape
completion + provably-dead nullCoalesce.offset guard removal.

Baseline regenerated; composer run analyse:tests green; 180 affected tests pass.

+++ ibl5/tests/Player/PlayerImageHelperTest.php
@@ -186,7 +186,10 @@ class PlayerImageHelperTest extends TestCase
      */
     public function testFloatPlayerIDIsConvertedCorrectly(): void
     {
-        $result = PlayerImageHelper::getImageUrl((int)123.7);
+        // Float input is intentional — this test verifies getImageUrl() itself
+        // converts a float ID. Pre-casting would gut it; any arg-type mismatch is
+        // a documented baseline defer.
+        $result = PlayerImageHelper::getImageUrl(123.7);
 
         // Should truncate to 123
         $this->assertStringContainsString('./images/player/123.jpg', $result);
